Moved add follower from AuthorRepository.cs to FollowerRepository.cs
