+++ src/Chirp.Core/IAuthorRepository.cs
@@ -7,5 +7,4 @@ public interface IAuthorRepository
     public AuthorDto GetAuthorByName(string authorName);
     public AuthorDto GetAuthorByEmail(string authorEmail);
     public void CreateAuthor(AuthorDto author);
-    public void AddFollower(string authorName, string followerName);
 }
