+++ src/Chirp.Core/IFollowerRepository.cs
@@ -1,0 +1,6 @@
+namespace Chirp.Core;
+
+public interface IFollowerRepository
+{
+    public void AddFollower(string authorName, string followerName);
+}
